fixed nsd generation
